NodeMCU : Use SoftwareSerial for communications
This avoids conflicts when uploading a new program to the NodeMCU and
thus removes the need to disconnect the serial line when uploading an
updated code.

It is still needed to disconnect the line when uploading to the Arduino
Nano but since the NodeMCU is the only one who needs to be updated to
change the automatic control parameters, the Arduino Nano shouldn't have
to be reprogrammed once everything is in place.

Signed-off-by: Maxime Chretien (MixLeNain) <mchretien@linuxmail.org>